+++ rebuild.sh
@@ -3,7 +3,7 @@
 # Example:
 # sh rebuild.sh /some/directory
 
-MAKEFILE="ucberkeley_cas-standalone.make"
+MAKEFILE="ssu_cas-standalone.make"
 AWK=${AWK:-awk}
 TAR=${TAR:-tar}
 LS=${LS:-ls}
@@ -16,10 +16,10 @@ read SELECTION
 
 if [ "$SELECTION" == "2" ];
 then
-  MAKEFILE="ucberkeley_cas-standalone-dev.make"
+  MAKEFILE="ssu_cas-standalone-dev.make"
 fi
 
-echo "Enter the full path at which you want to build ucberkeley_cas (default: /tmp): \c"
+echo "Enter the full path at which you want to build ssu_cas (default: /tmp): \c"
 read BUILD_DIR
 
 if [ -z "$BUILD_DIR" ];
@@ -36,34 +36,34 @@ fi
 echo "Building in $BUILD_DIR:\n"
 
 # remove any old builds
-if [ -d "$BUILD_DIR/ucberkeley_cas" ];
+if [ -d "$BUILD_DIR/ssu_cas" ];
 then
-  rm -rf $BUILD_DIR/ucberkeley_cas
+  rm -rf $BUILD_DIR/ssu_cas
 fi
 
-drush make -y --no-core --no-cache --contrib-destination=. $MAKEFILE $BUILD_DIR/build_ucberkeley_cas
-mv $BUILD_DIR/build_ucberkeley_cas/modules/* $BUILD_DIR
-mv $BUILD_DIR/cas* $BUILD_DIR/ucberkeley_cas/
-mv $BUILD_DIR/ldap $BUILD_DIR/ucberkeley_cas/
-mv $BUILD_DIR/build_ucberkeley_cas/libraries/phpcas/CAS* $BUILD_DIR/ucberkeley_cas/cas/CAS
-rm -rf $BUILD_DIR/build_ucberkeley_cas
-rm $BUILD_DIR/ucberkeley_cas/.gitignore
+drush make -y --no-core --no-cache --contrib-destination=. $MAKEFILE $BUILD_DIR/build_ssu_cas
+mv $BUILD_DIR/build_ssu_cas/modules/* $BUILD_DIR
+mv $BUILD_DIR/cas* $BUILD_DIR/ssu_cas/
+mv $BUILD_DIR/ldap $BUILD_DIR/ssu_cas/
+mv $BUILD_DIR/build_ssu_cas/libraries/phpcas/CAS* $BUILD_DIR/ssu_cas/cas/CAS
+rm -rf $BUILD_DIR/build_ssu_cas
+rm $BUILD_DIR/ssu_cas/.gitignore
 cd $BUILD_DIR
-VER=`$AWK -F = '/version =.*$/{gsub(/ /, "", $0); print $2}' ucberkeley_cas/ucberkeley_cas.info`
+VER=`$AWK -F = '/version =.*$/{gsub(/ /, "", $0); print $2}' ssu_cas/ssu_cas.info`
 echo ""
 while [[ ! "$CONFIRM" == "y" ]] && [[ ! "$CONFIRM" == "n" ]]; do
-  echo "Create this tarball: ucberkeley_cas-$VER.tar.gz? (y/n)"
+  echo "Create this tarball: ssu_cas-$VER.tar.gz? (y/n)"
   read CONFIRM
 done
 
 if [ "$CONFIRM" == "y" ];then
-  $TAR zcf ucberkeley_cas-$VER.tar.gz ucberkeley_cas
+  $TAR zcf ssu_cas-$VER.tar.gz ssu_cas
 else
   echo "Okay, no tarball."
 fi
 
 echo ""
-echo "$LS $BUILD_DIR/ucberkeley_cas* :"
+echo "$LS $BUILD_DIR/ssu_cas* :"
 echo ""
-$LS $BUILD_DIR/ucberkeley_cas*
+$LS $BUILD_DIR/ssu_cas*